edited two fiels
